+++ roles/roles_client.js
@@ -1,9 +1,10 @@
-/* global Meteor, Roles, Mongo */
+/* global Roles */
+import { Meteor } from 'meteor/meteor'
 
 /**
  * Provides functions related to user authorization. Compatible with built-in Meteor accounts packages.
  *
- * Roles are accessible throgh `Meteor.roles` collection and documents consist of:
+ * Roles are accessible through `Meteor.roles` collection and documents consist of:
  *  - `_id`: role name
  *  - `children`: list of subdocuments:
  *    - `_id`
@@ -14,7 +15,7 @@
  *
  * Example: `{_id: 'admin', children: [{_id: 'editor'}]}`
  *
- * The assignment of a role to a user is stored in a collection, accessible through `Meteor.roleAssignment`.
+ * The assignment of a role to a user is stored in a collection, accessible through `RoleAssignmentCollection`.
  * It's documents consist of
  *  - `_id`: Internal MongoDB id
  *  - `role`: A role object which got assigned. Usually only contains the `_id` property
@@ -24,13 +25,6 @@
  *
  * @module Roles
  */
-if (!Meteor.roles) {
-  Meteor.roles = new Mongo.Collection('roles')
-}
-
-if (!Meteor.roleAssignment) {
-  Meteor.roleAssignment = new Mongo.Collection('role-assignment')
-}
 
 /**
  * @class Roles
@@ -103,7 +97,7 @@ Object.assign(Roles, {
       // For all roles who have it as a dependency ...
       roles = Roles._getParentRoleNames(Meteor.roles.findOne({ _id: roleName }))
 
-      Meteor.roles.find({ _id: { $in: roles } }).fetch().forEach(r => {
+      for (const r of Meteor.roles.find({ _id: { $in: roles } }).fetch()) {
         Meteor.roles.update({
           _id: r._id
         }, {
@@ -122,7 +116,7 @@ Object.assign(Roles, {
             inheritedRoles: [r._id, ...inheritedRoles].map(r2 => ({ _id: r2 }))
           }
         }, { multi: true })
-      })
+      }
     } while (roles.length > 0)
 
     // And finally remove the role itself
@@ -203,9 +197,9 @@ Object.assign(Roles, {
     // ensure arrays
     if (!Array.isArray(rolesNames)) rolesNames = [rolesNames]
 
-    rolesNames.forEach(function (roleName) {
+    for (const roleName of rolesNames) {
       Roles._addRoleToParent(roleName, parentName)
-    })
+    }
   },
 
   /**
@@ -272,9 +266,9 @@ Object.assign(Roles, {
     // ensure arrays
     if (!Array.isArray(rolesNames)) rolesNames = [rolesNames]
 
-    rolesNames.forEach(function (roleName) {
+    for (const roleName of rolesNames) {
       Roles._removeRoleFromParent(roleName, parentName)
-    })
+    }
   },
 
   /**
@@ -313,7 +307,7 @@ Object.assign(Roles, {
     // For all roles who have had it as a dependency ...
     const roles = [...Roles._getParentRoleNames(Meteor.roles.findOne({ _id: parentName })), parentName]
 
-    Meteor.roles.find({ _id: { $in: roles } }).fetch().forEach(r => {
+    for (const r of Meteor.roles.find({ _id: { $in: roles } }).fetch()) {
       const inheritedRoles = Roles._getInheritedRoleNames(Meteor.roles.findOne({ _id: r._id }))
       Meteor.roleAssignment.update({
         'role._id': r._id,
@@ -323,7 +317,7 @@ Object.assign(Roles, {
           inheritedRoles: [r._id, ...inheritedRoles].map(r2 => ({ _id: r2 }))
         }
       }, { multi: true })
-    })
+    }
   },
 
   /**
@@ -365,17 +359,17 @@ Object.assign(Roles, {
       ifExists: false
     }, options)
 
-    users.forEach(function (user) {
+    for (const user of users) {
       if (typeof user === 'object') {
         id = user._id
       } else {
         id = user
       }
 
-      roles.forEach(function (role) {
+      for (const role of roles) {
         Roles._addUserToRole(id, role, options)
-      })
-    })
+      }
+    }
   },
 
   /**
@@ -419,7 +413,7 @@ Object.assign(Roles, {
       anyScope: false
     }, options)
 
-    users.forEach(function (user) {
+    for (const user of users) {
       if (typeof user === 'object') {
         id = user._id
       } else {
@@ -434,10 +428,10 @@ Object.assign(Roles, {
       Meteor.roleAssignment.remove(selector)
 
       // and then add all
-      roles.forEach(function (role) {
+      for (const role of roles) {
         Roles._addUserToRole(id, role, options)
-      })
-    })
+      }
+    }
   },
 
   /**
@@ -512,11 +506,11 @@ Object.assign(Roles, {
 
     const parentRoles = new Set([role._id])
 
-    parentRoles.forEach(roleName => {
+    for (const roleName of parentRoles) {
       Meteor.roles.find({ 'children._id': roleName }).fetch().forEach(parentRole => {
         parentRoles.add(parentRole._id)
       })
-    })
+    }
 
     parentRoles.delete(role._id)
 
@@ -538,14 +532,14 @@ Object.assign(Roles, {
     const inheritedRoles = new Set()
     const nestedRoles = new Set([role])
 
-    nestedRoles.forEach(r => {
+    for (const r of nestedRoles) {
       const roles = Meteor.roles.find({ _id: { $in: r.children.map(r => r._id) } }, { fields: { children: 1 } }).fetch()
 
-      roles.forEach(r2 => {
+      for (const r2 of roles) {
         inheritedRoles.add(r2._id)
         nestedRoles.add(r2)
-      })
-    })
+      }
+    }
 
     return [...inheritedRoles]
   },
@@ -580,10 +574,10 @@ Object.assign(Roles, {
 
     Roles._checkScopeName(options.scope)
 
-    users.forEach(function (user) {
-      if (!user) return
+    for (const user of users) {
+      if (!user) continue
 
-      roles.forEach(function (role) {
+      for (const role of roles) {
         let id
         if (typeof user === 'object') {
           id = user._id
@@ -592,8 +586,8 @@ Object.assign(Roles, {
         }
 
         Roles._removeUserFromRole(id, role, options)
-      })
-    })
+      }
+    }
   },
 
   /**
@@ -775,7 +769,7 @@ Object.assign(Roles, {
    *
    * @method getAllRoles
    * @param {Object} queryOptions Options which are passed directly
-   *                                through to `Meteor.roles.find(query, options)`.
+   *                                through to `RolesCollection.find(query, options)`.
    * @return {Cursor} Cursor of existing roles.
    * @static
    */
@@ -830,7 +824,7 @@ Object.assign(Roles, {
    *     roles will also be checked
    *   - `anyScope`: if set, role can be in any scope (`scope` option is ignored)
    *   - `queryOptions`: options which are passed directly
-   *     through to `Meteor.roleAssignment.find(query, options)`
+   *     through to `RoleAssignmentCollection.find(query, options)`
    *
    * Alternatively, it can be a scope name string.
    * @return {Cursor} Cursor of user assignments for roles.
@@ -860,7 +854,7 @@ Object.assign(Roles, {
    *
    * Alternatively, it can be a scope name string.
    * @param {Object} [filter] Options which are passed directly
-   *                                through to `Meteor.roleAssignment.find(query, options)`
+   *                                through to `RoleAssignmentCollection.find(query, options)`
    * @return {Object} Cursor to the assignment documents
    * @private
    * @static
@@ -1097,8 +1091,12 @@ Object.assign(Roles, {
   _checkScopeName: function (scopeName) {
     if (scopeName === null) return
 
-    if (!scopeName || typeof scopeName !== 'string' || scopeName.trim() !== scopeName) {
-      throw new Error('Invalid scope name \'' + scopeName + '\'.')
+    if (
+      !scopeName ||
+      typeof scopeName !== 'string' ||
+      scopeName.trim() !== scopeName
+    ) {
+      throw new Error(`Invalid scope name '${scopeName}'.`)
     }
   }
 })